+++ src/avt/Pipeline/Data/avtWedgeExtractor.C
@@ -230,7 +230,7 @@ avtWedgeExtractor::Extract(const avtWedge &wdg)
     int maxx_iter = (maxx > restrictedMaxWidth ? restrictedMaxWidth : maxx);
     for (int xi = minx_iter ; xi <= maxx_iter ; xi++)
     {
-        float x = XFromIndex(xi);
+        double x = XFromIndex(xi);
 
         int   triIndex = IndexToTriangulationTable(wdg.pts, 6, x);
 
@@ -247,7 +247,7 @@ avtWedgeExtractor::Extract(const avtWedge &wdg)
             // triangle intersects and then interpolating along that edge
             // three times to form the triangle.
             //
-            float y[3], z[3], v[3][AVT_VARIABLE_LIMIT];
+            double y[3], z[3], v[3][AVT_VARIABLE_LIMIT];
             for (int tri_vertex = 0 ; tri_vertex < 3 ; tri_vertex++)
             {
                 int wdg_vertex1 = verticesFromEdges[triCase[tri_vertex]][0];
